backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.8.1
pandas==2.2.3
numpy==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
supabase==2.13.0
google-genai==1.10.0
requests==2.32.3
aiofiles==24.1.0
openpyxl==3.1.5
xlrd==2.0.1
python-json-logger==3.3.0
prometheus-client==0.21.1
psutil==6.1.1
cachetools==5.5.2
slowapi==0.1.9
PyJWT==2.10.1
celery==5.4.0
redis==5.2.1
# Dev / testing
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
